+++ src/paimon/CMakeLists.txt
@@ -369,7 +369,7 @@ set(PAIMON_CORE_SRCS
     core/realtime/arrow_mem_indexer.cpp
     core/realtime/arrow_mem_indexer_factory.cpp
     core/realtime/realtime_append_only_writer.cpp
-    core/realtime/realtime_context.cpp
+    core/realtime/realtime_context_impl.cpp
     core/postpone/postpone_bucket_writer.cpp
     core/schema/arrow_schema_validator.cpp
     core/schema/schema_manager.cpp
@@ -755,6 +755,7 @@ if(PAIMON_BUILD_TESTS)
                     core/memory/writer_memory_manager_test.cpp
                     core/realtime/arrow_mem_indexer_test.cpp
                     core/realtime/realtime_context_test.cpp
+                    core/realtime/realtime_reader_test.cpp
                     core/mergetree/levels_test.cpp
                     core/mergetree/lookup_file_test.cpp
                     core/mergetree/lookup_levels_test.cpp
